Fix tests: create users in setUp, update authenticate, resolve throttling issues

+++ cinema/tests/test_movie_api.py
@@ -1,10 +1,11 @@
 import tempfile
 import os
-
 from PIL import Image
 from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.urls import reverse
+from django.utils import timezone
+from datetime import datetime
 
 from rest_framework.test import APIClient
 from rest_framework import status
@@ -22,23 +23,18 @@ def sample_movie(**params):
         "duration": 90,
     }
     defaults.update(params)
-
     return Movie.objects.create(**defaults)
 
 
 def sample_genre(**params):
-    defaults = {
-        "name": "Drama",
-    }
+    defaults = {"name": "Drama"}
     defaults.update(params)
-
     return Genre.objects.create(**defaults)
 
 
 def sample_actor(**params):
     defaults = {"first_name": "George", "last_name": "Clooney"}
     defaults.update(params)
-
     return Actor.objects.create(**defaults)
 
 
@@ -46,19 +42,16 @@ def sample_movie_session(**params):
     cinema_hall = CinemaHall.objects.create(
         name="Blue", rows=20, seats_in_row=20
     )
-
     defaults = {
-        "show_time": "2022-06-02 14:00:00",
+        "show_time": timezone.make_aware(datetime(2022, 6, 2, 14, 0, 0)),
         "movie": None,
         "cinema_hall": cinema_hall,
     }
     defaults.update(params)
-
     return MovieSession.objects.create(**defaults)
 
 
 def image_upload_url(movie_id):
-    """Return URL for recipe image upload"""
     return reverse("cinema:movie-upload-image", args=[movie_id])
 
 
@@ -69,20 +62,31 @@ def detail_url(movie_id):
 class MovieImageUploadTests(TestCase):
     def setUp(self):
         self.client = APIClient()
-        self.user = get_user_model().objects.create_superuser(
-            "admin@myproject.com", "password"
+        # створюємо суперкористувача у тестовій базі
+        self.admin = get_user_model().objects.create_superuser(
+            email="admin@myproject.com",
+            password="AdminPass123"
         )
-        self.client.force_authenticate(self.user)
+        # отримуємо JWT токен
+        response = self.client.post("/api/token/", {
+            "email": "admin@myproject.com",
+            "password": "AdminPass123"
+        })
+        self.assertEqual(response.status_code, 200, msg=response.data)
+        token = response.data.get("access")
+        self.assertIsNotNone(token, msg=response.data)
+        self.client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
+
         self.movie = sample_movie()
         self.genre = sample_genre()
         self.actor = sample_actor()
         self.movie_session = sample_movie_session(movie=self.movie)
 
     def tearDown(self):
-        self.movie.image.delete()
+        if self.movie.image:
+            self.movie.image.delete()
 
     def test_upload_image_to_movie(self):
-        """Test uploading an image to movie"""
         url = image_upload_url(self.movie.id)
         with tempfile.NamedTemporaryFile(suffix=".jpg") as ntf:
             img = Image.new("RGB", (10, 10))
@@ -96,10 +100,8 @@ class MovieImageUploadTests(TestCase):
         self.assertTrue(os.path.exists(self.movie.image.path))
 
     def test_upload_image_bad_request(self):
-        """Test uploading an invalid image"""
         url = image_upload_url(self.movie.id)
         res = self.client.post(url, {"image": "not image"}, format="multipart")
-
         self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
 
     def test_post_image_to_movie_list(self):
@@ -114,13 +116,12 @@ class MovieImageUploadTests(TestCase):
                     "title": "Title",
                     "description": "Description",
                     "duration": 90,
-                    "genres": [1],
-                    "actors": [1],
+                    "genres": [self.genre.id],
+                    "actors": [self.actor.id],
                     "image": ntf,
                 },
                 format="multipart",
             )
-
         self.assertEqual(res.status_code, status.HTTP_201_CREATED)
         movie = Movie.objects.get(title="Title")
         self.assertFalse(movie.image)
@@ -133,7 +134,6 @@ class MovieImageUploadTests(TestCase):
             ntf.seek(0)
             self.client.post(url, {"image": ntf}, format="multipart")
         res = self.client.get(detail_url(self.movie.id))
-
         self.assertIn("image", res.data)
 
     def test_image_url_is_shown_on_movie_list(self):
@@ -144,7 +144,6 @@ class MovieImageUploadTests(TestCase):
             ntf.seek(0)
             self.client.post(url, {"image": ntf}, format="multipart")
         res = self.client.get(MOVIE_URL)
-
         self.assertIn("image", res.data[0].keys())
 
     def test_image_url_is_shown_on_movie_session_detail(self):
@@ -155,5 +154,4 @@ class MovieImageUploadTests(TestCase):
             ntf.seek(0)
             self.client.post(url, {"image": ntf}, format="multipart")
         res = self.client.get(MOVIE_SESSION_URL)
-
         self.assertIn("movie_image", res.data[0].keys())
